VM2GKE/codelab/requirements.txt
cherrypy
sqlalchemy
pyyaml
pymysql